requirements.txt
streamlit==1.41.1
langchain==0.3.17
langchain-community==0.3.16
langchain-openai==0.3.3
openai==1.61.0
python-dotenv==1.0.0
pypdf==5.2.0
tiktoken==0.8.0
python-magic==0.4.27
python-magic-bin==0.4.14
langchain-text-splitters==0.3.5
supabase==2.12.0
